% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\docType{data}
\name{dict}
\alias{dict}
\title{Dictionary of countries}
\format{An object of class \code{data.frame} with 328 rows and 2 columns.}
\usage{
dict
}
\description{
Dictionary of countries
}
\keyword{datasets}
