% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getc.R
\name{getc}
\alias{getc}
\title{Return the correct country name}
\usage{
getc(country)
}
\arguments{
\item{country}{country name to be checked}
}
\value{
correct country name
}
\description{
Return the correct country name
}
\examples{
getc("USA")
}
