% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Imputation by Interpolation and Extrapolation}
\usage{
impute(x)
}
\arguments{
\item{x}{EPI data frame with -9999/-8888/-7777 codes}
}
\value{
EPI data frame with missing values imputed by interpolation and
exterpolation and -8888 preserved
}
\description{
Imputes missing values by interpolation and extrapolation
}
\details{
EPI data frame follows the 2018 Masterfile convention with the first
3 columns being [code, iso, country] and the following columns being
VARIABLE.STEP.YEAR (e.g. MPA.trf.2005)
}
