% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Log or Log-inverse transformation}
\usage{
logit(x, shift, inverse = FALSE, maxval = 100)
}
\arguments{
\item{x}{EPI data frame to be log tranformed. There cannot be negative values
or missing values coded as -9999/-8888.}

\item{shift}{shift size, zero by default. Can be set to "auto", then it will
be derived from the data}

\item{inverse}{loginverse or not}

\item{maxval}{value to subtract data values from}
}
\value{
log-transformed EPI data frame
}
\description{
Log tranforms the data with an appropriate adjustments
}
\details{
EPI data frame follows the 2018 Masterfile convention with the first
3 columns being [code, iso, country] and the following columns being
VARIABLE.STEP.YEAR (e.g. MPA.trf.2005)
}
