% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.na.R
\name{make.na}
\alias{make.na}
\title{Make NA}
\usage{
make.na(folder)
}
\arguments{
\item{folder}{folder with .csv files}
}
\description{
Converts EPI missing value codes to NA for all files in the given folder.
Resulting files are put in the same folder with '_na' added to their original
names.
}
