% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterfile.R
\docType{data}
\name{MasterFile}
\alias{MasterFile}
\title{Master file}
\format{An object of class \code{data.frame} with 236 rows and 3 columns.}
\usage{
MasterFile
}
\description{
Master file
}
\keyword{datasets}
