% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.csv.epi.R
\name{read.csv.epi}
\alias{read.csv.epi}
\title{Data Input}
\usage{
read.csv.epi(file, root = NA, year = 2016, allow.na = TRUE)
}
\arguments{
\item{file}{file that contains the data}

\item{root}{root variable name to check that the data file has correct column
names}

\item{year}{most recent year. If there is a year in the data file that is
after \code{year}, warning will be raised.}

\item{allow.na}{if FALSE, warning will be raised if there are NAs in the file}
}
\value{
data frame read-in as by \code{read.csv}
}
\description{
In addition to \code{read.csv} functionality, this function checks whether
the data file satisfies the EPI rules:
1) The first three columns must be 'code', 'iso' and 'country' following
   the Masterfile format.
2) All following columns must follow the convention 'NAME.YYYY' where NAME is
   the variable name and YYYY the year.
3) NAME must be consistent and same with what is in the file name.
}
