% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resize.R
\name{resize}
\alias{resize}
\title{Resizing the Data Frame}
\usage{
resize(x, start = 2000, end = 2017)
}
\arguments{
\item{x}{EPI data frame to be resized}

\item{start}{start year}

\item{end}{end year}
}
\value{
resized EPI data frame
}
\description{
Blows out or cuts the data frame with the given start and
end year. Nonexistent year columns are filled with NA. STEP TLA in column
names is changed to rsz.
}
\details{
EPI data frame follows the 2018 Masterfile convention with the first
3 columns being [code, iso, country] and the following columns being
VARIABLE.STEP.YEAR (e.g. MPA.trf.2005)
}
