% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totarget.R
\name{totarget}
\alias{totarget}
\title{Scales the data with target values}
\usage{
totarget(x, low, high, polarity = "negative")
}
\arguments{
\item{x}{data to be scaled with target values}

\item{low}{low target value}

\item{high}{high target value}

\item{polarity}{polarity, either positive or negative}
}
\value{
scaled data
}
\description{
For positive polarity, high target is considered good and low target bad.
Values under bad target are converted to 0 and values over good target are
converted to 100. All other values are scaled in between 0 and 100. The
opposite is performed for negative polarity.
}
